Repair!

+++ Seminar2/Homework2/src/Coordonates.java
@@ -1,7 +1,7 @@
 public class Coordonates {
-    Float x;
-    Float y;
-    public Coordonates(Float x1,Float y1){
+    Double x;
+    Double y;
+    public Coordonates(Double x1,Double y1){
         x=x1;
         y=y1;
     }
